+++ src/main/java/Backend/ms_clasificator/Services/PatientService.java
@@ -3,7 +3,9 @@ package Backend.ms_clasificator.Services;
 import Backend.ms_clasificator.DTOs.Patient.PatientCreateDTO;
 import Backend.ms_clasificator.DTOs.Response.ApiResponse;
 import Backend.ms_clasificator.Mappers.PatientMappers.PatientMapper;
+import Backend.ms_clasificator.Models.MedicalImg;
 import Backend.ms_clasificator.Models.Patient;
+import Backend.ms_clasificator.Repositories.MedicalImgRepository;
 import Backend.ms_clasificator.Repositories.PatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -21,6 +23,9 @@ public class PatientService {
 
     @Autowired
     private PatientMapper patientMapper;
+
+    @Autowired
+    private MedicalImgRepository medicalImgRepository;
 
     /**
      * Obtener todos los pacientes
@@ -146,11 +151,18 @@ public class PatientService {
             Patient patient = patientRepository.findById(id)
                     .orElseThrow(() -> new IllegalArgumentException("Paciente no encontrado con ID: " + id));
 
+            // Validamos que el paciente no tenga imagenes asociadas
+            List<MedicalImg> medicalImgs = this.medicalImgRepository.findByPatientId(id);
+            if (!medicalImgs.isEmpty()) {
+                return ApiResponse.error("No se puede eliminar el paciente porque tiene imagenes medicas asociadas");
+            }
+
+
             patientRepository.delete(patient);
             return ApiResponse.success("Paciente eliminado exitosamente");
 
         } catch (DataIntegrityViolationException ex) {
-            return ApiResponse.error("No se puede eliminar el paciente porque tiene imagenes medicas asociadas");
+            return ApiResponse.error("Violacion a integridad de la base de datos: " + ex.getMessage());
         } catch (IllegalArgumentException ex) {
             return ApiResponse.error(ex.getMessage());
         } catch (Exception ex) {
